+++ app/components/events-page/edit-event.tsx
@@ -14,6 +14,7 @@ import { useRouter } from 'next/navigation';
 import RegistrationsModal from './registrations-modal';
 import ToggleSwitch from '../toggle-button';
 import { createPortal } from 'react-dom';
+import EventEmailSendingModal from './email-sending-modal';
 
 
 const MAX_POSTGRES_STRING_LENGTH = 255;
@@ -23,6 +24,7 @@ export default function EditEventComponent({ eventProp, onClose }: EditEventProp
 	const [viewRegistrationsModal, setViewRegistrationsModal] = useState(false)
 	const [organisers, setOrganisers] = useState([eventProp.organiser]);
 	const [registrations, setRegistrations] = useState<Registrations[]>([])
+	const [viewSendEmailsModal, setViewSendEmailsModal] = useState<boolean>(false); // this should be in conflict with viewRegistrations, only one can be true at the same time
 
 	const { register, handleSubmit, formState: { errors, isValid }, setValue, watch } = useForm<FormData>({
 		mode: 'onChange',
@@ -57,6 +59,18 @@ export default function EditEventComponent({ eventProp, onClose }: EditEventProp
 		}
 
 		setViewRegistrationsModal(true)
+		// just in case
+		setViewSendEmailsModal(false)
+	}
+
+	useEffect(() => {
+		console.log("changed");
+	}, [viewSendEmailsModal, viewRegistrationsModal])
+
+	const viewSendEmails = () => {
+		setViewSendEmailsModal(true)
+		// just in case
+		setViewRegistrationsModal(false)
 	}
 
 	const fetchOrganisersData = async () => {
@@ -138,23 +152,25 @@ export default function EditEventComponent({ eventProp, onClose }: EditEventProp
 	}, []);
 
 	// Disable background scroll and handle outside click detection
-	useEffect(() => {
-		// Prevent background scrolling
-		document.body.style.overflow = 'hidden';
+	// but this closes the modal when it is covered, say opening the registration
+	// so actually, I dont think we need this
+	// useEffect(() => {
+	// 	// Prevent background scrolling
+	// 	document.body.style.overflow = 'hidden';
 
-		const handleClickOutside = (event: MouseEvent) => {
-			if (modalRef.current && !modalRef.current.contains(event.target as Node)) {
-				onClose();
-			}
-		};
+	// 	const handleClickOutside = (event: MouseEvent) => {
+	// 		if (modalRef.current && !modalRef.current.contains(event.target as Node)) {
+	// 			onClose();
+	// 		}
+	// 	};
 
-		document.addEventListener('mousedown', handleClickOutside);
+	// 	document.addEventListener('mousedown', handleClickOutside);
 
-		return () => {
-			document.body.style.overflow = '';
-			document.removeEventListener('mousedown', handleClickOutside);
-		};
-	}, [onClose]);
+	// 	return () => {
+	// 		document.body.style.overflow = '';
+	// 		document.removeEventListener('mousedown', handleClickOutside);
+	// 	};
+	// }, [onClose]);
 
 	const onSubmit = async (data: FormData) => {
 		const toastId = toast.loading('Updating event....')
@@ -557,6 +573,14 @@ export default function EditEventComponent({ eventProp, onClose }: EditEventProp
 									View Registrations
 								</Button>
 								<Button
+									variant="outline"
+									size="sm"
+									className="px-10 bg-purple-400 text-gray-950 md:text-xl"
+									onClick={viewSendEmails}
+								>
+									Send Emails to attendee
+								</Button>
+								<Button
 									className="px-10 md:text-xl border-black"
 									variant="outline"
 									size="sm"
@@ -587,6 +611,7 @@ export default function EditEventComponent({ eventProp, onClose }: EditEventProp
 	
 						{/* Modal for Registrations */}
 						{viewRegistrationsModal && <RegistrationsModal registrations={registrations} onClose={closeModal} />}
+						{viewSendEmailsModal && <EventEmailSendingModal onClose={() => {console.log("closed");setViewSendEmailsModal(false)}} event={eventProp}/>}
 					</div>
 				</div>
 			</div>