+++ app/societies/(overview)/page.tsx
@@ -84,7 +84,9 @@ export default function SocietyPage() {
 			)} */}
 
 			{/* Partners List */}
-			<div className=" w-full mt-3 grid partner-grid gap-8 ">
+			{/* notice that partner grid is defined in global.css */}
+			<div className="relative w-full mt-3 grid partner-grid gap-8">
+				{/* <Partners filteredPartners={[...(partner.map(it => {it.logo = null;return it})), ...filteredPartners]} skeleton={loading}/> only for testing */}
 				<Partners filteredPartners={filteredPartners} skeleton={loading}/>
 			</div>